Update order from cart or buy now

+++ src/services/order.service.js
@@ -4,33 +4,63 @@ const AppError = require('../utils/appError')
 class OrderService {
 
     // Tạo đơn hàng từ giỏ hàng
-    async createOrder(userId) {
+    // Thay vì chỉ nhận userId, ta nhận thêm directItem nếu là "Mua ngay"
+    async createOrder(userId, directItem = null) {
         return await prisma.$transaction(async (tx) => {
 
-            // B1 - Lấy giỏ hàng
-            const cart = await tx.carts.findUnique({
-                where: {user_id: userId},
-                include: {
-                    cart_items: {
-                        include: {products: true}
+            let itemsToOrder = [];
+
+            // TRƯỜNG HỢP 1: NGƯỜI DÙNG BẤM "MUA NGAY" TRÊN FRONT-END
+            if (directItem) {
+                // directItem dạng: { product_id: 29, quantity: 1 }
+                const product = await tx.products.findUnique({
+                    where: { id: directItem.product_id }
+                });
+
+                if (!product) {
+                    throw new AppError('Sản phẩm không tồn tại', 404);
+                }
+                if (product.stock < directItem.quantity) {
+                    throw new AppError(`${product.name} chỉ còn ${product.stock} sản phẩm`, 400);
+                }
+
+                // Gán thông tin cấu trúc giống hệt như lấy từ cart_items để dùng chung logic bên dưới
+                itemsToOrder.push({
+                    product_id: product.id,
+                    quantity: directItem.quantity,
+                    price: product.price,
+                    products: product
+                });
+
+                // TRƯỜNG HỢP 2: CHECKOUT TỪ GIỎ HÀNG (Code cũ của bạn)
+            } else {
+                const cart = await tx.carts.findUnique({
+                    where: { user_id: userId },
+                    include: {
+                        cart_items: {
+                            include: { products: true }
+                        }
+                    }
+                });
+
+                if (!cart || cart.cart_items.length === 0) {
+                    throw new AppError('Giỏ hàng trống', 400);
+                }
+
+                // Kiểm tra stock từng sản phẩm trong giỏ
+                for (const item of cart.cart_items) {
+                    if (item.products.stock < item.quantity) {
+                        throw new AppError(`${item.products.name} chỉ còn ${item.products.stock} sản phẩm`, 400);
                     }
                 }
-            })
-            if(!cart || cart.cart_items.length === 0){
-                throw new Error('Giỏ hàng trống')
+
+                itemsToOrder = cart.cart_items;
             }
 
-            // B2 - Kiểm tra stock từng sản phẩm
-            for(const item of cart.cart_items) {
-                if(item.products.stock < item.quantity){
-                    throw new Error(`${item.products.name} chỉ còn ${item.products.stock} sản phẩm`)
-                }
-            }
-
-            // B3 - Tính tổng tiền
-            const total = cart.cart_items.reduce((sum, item) => {
-                return sum + Number(item.products.price) * item.quantity
-            }, 0)
+            // B3 - Tính tổng tiền (Dùng chung cho cả 2 trường hợp)
+            const total = itemsToOrder.reduce((sum, item) => {
+                return sum + Number(item.products.price) * item.quantity;
+            }, 0);
 
             // B4 - Tạo order
             const order = await tx.orders.create({
@@ -39,81 +69,82 @@ class OrderService {
                     total,
                     status: 'pending',
                     order_items: {
-                        create: cart.cart_items.map(item => ({
+                        create: itemsToOrder.map(item => ({
                             product_id: item.product_id,
                             quantity: item.quantity,
-                            price: item.products.price
+                            price: item.products.price // Chụp lại giá tại thời điểm mua
                         }))
                     }
                 },
                 include: {
                     order_items: {
                         include: {
-                            products: {
-                                select: {id: true, name: true}
-                            }
+                            products: { select: { id: true, name: true } }
                         }
                     }
                 }
-            })
+            });
 
-            // B6 - Trừ stock
-            for(const item of cart.cart_items){
+            // B6 - Trừ stock sản phẩm
+            for (const item of itemsToOrder) {
                 await tx.products.update({
-                    where: {id: item.product_id},
-                    data: {stock: { decrement: item.quantity}}
-                })
+                    where: { id: item.product_id },
+                    data: { stock: { decrement: item.quantity } }
+                });
             }
 
-            // B7 - Xóa giỏ hàng
-            await tx.cart_items.deleteMany({
-                where: {cart_id: cart.id}
-            })
+            // B7 - Xóa giỏ hàng (CHỈ thực hiện nếu user thanh toán từ nút trong Giỏ Hàng)
+            if (!directItem) {
+                const cart = await tx.carts.findUnique({ where: { user_id: userId } });
+                await tx.cart_items.deleteMany({
+                    where: { cart_id: cart.id }
+                });
+            }
 
-            return order
-        })
+            return order;
+        });
     }
 
     // Lấy lịch sử đơn hàng của user
-    async getMyOrders(userId){
+    async getMyOrders(userId) {
         return await prisma.orders.findMany({
-            where: {user_id: userId},
+            where: { user_id: userId },
             include: {
                 order_items: {
                     include: {
                         products: {
-                            select: {id: true, name: true, price: true}
+                            select: { id: true, name: true, price: true }
                         }
                     }
                 }
             },
-            orderBy: {id: 'desc'}
+            orderBy: { id: 'desc' }
         })
     }
 
     // Lấy chi tiết 1 đơn hàng
-    async getOrderById(orderId, userId, role){ 
+    async getOrderById(orderId, userId, role) {
         const order = await prisma.orders.findUnique({
-            where: {id: orderId},
+            where: { id: orderId },
             include: {
                 users: {
-                    select: {id: true, name: true, email: true}
+                    select: { id: true, name: true, email: true }
                 },
                 order_items: {
                     include: {
                         products: {
-                            select: {id: true, name: true, price: true}
+                            select: { id: true, name: true, price: true }
                         }
                     }
                 }
             }
         })
-        if(!order){
+        if (!order) {
             throw new AppError("Không tìm thấy đơn hàng", 404)
         }
 
         // User chỉ xem đơn của mình
-        if(role !== 'admin' && order.user_id !== userId){
+        if (role !== 'admin' && order.user_id !== userId) {
             throw new AppError("Không có quyền xem đơn hàng này", 403)
         }
 
@@ -121,89 +152,89 @@ class OrderService {
     }
 
     // Admin lấy tất cả đơn hàng
-    async getAllOrders(query){
-        const {status, page = 1, limit = 10} = query
+    async getAllOrders(query) {
+        const { status, page = 1, limit = 10 } = query
 
-        const where = status ? {status} : {}
+        const where = status ? { status } : {}
 
         const [orders, total] = await Promise.all([
             prisma.orders.findMany({
                 where,
                 include: {
                     users: {
-                        select: {id: true, name: true, email: true}
+                        select: { id: true, name: true, email: true }
                     },
                     order_items: {
-                        include:{
+                        include: {
                             products: {
-                                select: {id: true, name: true}
+                                select: { id: true, name: true }
                             }
                         }
                     }
                 },
-                orderBy: {id: 'desc'},
+                orderBy: { id: 'desc' },
                 skip: (page - 1) * limit,
                 take: parseInt(limit)
             }),
-            prisma.orders.count({where})
+            prisma.orders.count({ where })
         ])
 
         return {
             data: orders,
             total,
             page: parseInt(page),
-            totalPages: Math.ceil(total/limit)
+            totalPages: Math.ceil(total / limit)
         }
     }
 
     // Cập nhật status - Admin
-    async updateStatus (orderId, status) {
+    async updateStatus(orderId, status) {
         const order = await prisma.orders.findUnique({
-            where: {id: orderId}
+            where: { id: orderId }
         })
-        if(!order){
+        if (!order) {
             throw new AppError('Không tìm thấy đơn hàng', 404)
         }
 
         // Không thể thay đổi đơn đã cancell/deli
-        if(['cancelled', 'delivered'].includes(order.status)){
+        if (['cancelled', 'delivered'].includes(order.status)) {
             throw new AppError(`Không thể thay đổi đơn hàng đã ${order.status}`, 400)
         }
 
         return await prisma.orders.update({
-            where: {id: orderId},
-            data: {status}
+            where: { id: orderId },
+            data: { status }
         })
     }
 
     // Hủy đơn - User
-    async cancelOrder(orderId, userId){
+    async cancelOrder(orderId, userId) {
         return await prisma.$transaction(async (tx) => {
             const order = await tx.orders.findUnique({
-                where: {id: orderId},
-                include: {order_items: true}
+                where: { id: orderId },
+                include: { order_items: true }
             })
-            if(!order){
+            if (!order) {
                 throw new AppError("Không tìm thấy đơn hàng", 404)
             }
-            if(order.user_id !== userId){
+            if (order.user_id !== userId) {
                 throw new AppError("Không có quyền hủy đơn hàng này", 403)
             }
-            if(order.status !== 'pending'){
+            if (order.status !== 'pending') {
                 throw new AppError("Chỉ hủy được đơn hàng đang chờ xử lý", 400)
             }
 
             // Hoàn lại stock
-            for(const item of order.order_items){
+            for (const item of order.order_items) {
                 await tx.products.update({
-                    where: {id: item.product_id},
-                    data: {stock: {increment: item.quantity}}
+                    where: { id: item.product_id },
+                    data: { stock: { increment: item.quantity } }
                 })
             }
 
             return await tx.orders.update({
-                where: {id: orderId},
-                data: {status: 'cancelled'}
+                where: { id: orderId },
+                data: { status: 'cancelled' }
             })
         })
     }
